pyproject.toml
[project]
name = "streamlit-hello-world"
version = "0.1.0"
description = "Simple Streamlit app with UV and Docker"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.39.0",
    "opencv-python-headless>=4.8.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"